swincyst/requirements.txt
fastapi==0.115.8
matplotlib==3.10.0
numpy==2.2.3
Pillow==11.1.0
scikit_learn==1.6.1
seaborn==0.13.2
timm==1.0.15
torch
torchvision==0.21.0
tqdm==4.67.1
uvicorn==0.34.0
python-multipart==0.0.9